@@ -30,7 +30,7 @@
         "@skbkontur-moira/codemirror-metricsql": "0.0.1-dev.1809240003",
         "@skbkontur/react-icons": "5.2.9",
         "@skbkontur/react-stack-layout": "1.0.3",
-        "@skbkontur/react-ui": "4.16.0",
+        "@skbkontur/react-ui": "5.1.3",
         "@skbkontur/react-ui-validations": "1.13.0",
         "@tanstack/react-table": "8.17.3",
         "@types/codemirror": "5.60.1",
@@ -47,17 +47,16 @@
         "npm-run-all": "4.1.5",
         "numeral": "2.0.6",
         "query-string": "6.4.0",
-        "react": "16.14.0",
-        "react-animate-height": "2.0.23",
+        "react": "18.3.1",
+        "react-animate-height": "3.2.3",
         "react-chartjs-2": "5.2.0",
-        "react-dom": "16.14.0",
+        "react-dom": "18.3.1",
         "react-favicon": "2.0.7",
         "react-foco": "1.2.0",
-        "react-hot-loader": "4.13.0",
         "react-markdown": "8.0.7",
         "react-redux": "8.1.3",
         "react-router": "7.5.2",
-        "react-router-dom": "5.1.2",
+        "react-router-dom": "7.5.2",
         "react-virtualized-auto-sizer": "1.0.20",
         "reflect-metadata": "0.1.13",
         "remark-breaks": "4.0.0",
@@ -68,7 +67,6 @@
         "@babel/core": "7.23.3",
         "@babel/preset-env": "7.23.3",
         "@babel/preset-typescript": "7.13.0",
-        "@hot-loader/react-dom": "16.13.0",
         "@lezer/generator": "1.0.0",
         "@playwright/test": "1.40.1",
         "@storybook/addon-actions": "7.6.17",
@@ -82,10 +80,9 @@
         "@types/jest": "26.0.9",
         "@types/lodash": "4.14.159",
         "@types/numeral": "0.0.28",
-        "@types/react": "16.9.43",
-        "@types/react-dom": "16.9.8",
-        "@types/react-router-dom": "5.1.5",
-        "@types/react-window": "1.8.5",
+        "@types/react": "18.3.1",
+        "@types/react-dom": "18.3.1",
+        "@types/react-window": "1.8.0",
         "@types/webpack-env": "1.18.4",
         "@typescript-eslint/eslint-plugin": "6.9.0",
         "@typescript-eslint/parser": "6.9.0",
@@ -112,7 +109,7 @@
         "mockdate": "3.0.2",
         "prettier": "2.0.5",
         "react-docgen-typescript-plugin": "0.6.0",
-        "react-window": "1.8.9",
+        "react-window": "1.8.0",
         "regenerator-runtime": "0.13.5",
         "storybook": "7.6.17",
         "style-loader": "3.3.3",
@@ -127,9 +124,6 @@
         "workbox-webpack-plugin": "7.1.0"
     },
     "resolutions": {
-        "@types/react-router": "5.1.18",
-        "@types/react": "16.9.43",
-        "@types/history": "4.7.11",
         "event-source-polyfill": "1.0.25",
         "es5-ext": "0.10.53"
     }
