@@ -36,7 +36,7 @@ import { LinkMenuItem } from "./Components/LinkMenuItem";
 import { ScheduleView } from "./Components/ScheduleView";
 import { ConfigState } from "../../store/selectors";
 import useConfirmModal from "../../hooks/useConfirmModal";
-import { useHistory } from "react-router";
+import { useNavigate } from "react-router";
 import { Flexbox } from "../Flexbox/FlexBox";
 import classNames from "classnames/bind";
 
@@ -55,13 +55,7 @@ interface IProps {
 }
 
 function maintenanceCaption(delta: number): React.ReactNode {
-    return (
-        <span>
-            <ClockIcon />
-            &nbsp;
-            {delta <= 0 ? "Maintenance" : humanizeDuration(delta)}
-        </span>
-    );
+    return <span>{delta <= 0 ? "Maintenance" : humanizeDuration(delta)}</span>;
 }
 
 export default function TriggerInfo({
@@ -92,7 +86,7 @@ export default function TriggerInfo({
     const { state, msg: exceptionMessage, maintenance, maintenance_info } = triggerState;
     const { config } = useAppSelector(ConfigState);
     const [ConfirmModal, setModalData] = useConfirmModal();
-    const history = useHistory();
+    const navigate = useNavigate();
 
     const availableClusters = config?.metric_source_clusters?.filter(
         (cluster) => cluster.trigger_source === triggerSource
@@ -136,9 +130,13 @@ export default function TriggerInfo({
                     {name != null && name !== "" ? name : "[No name]"}
                 </h1>
                 <div className={cn("controls")}>
-                    <Link data-tid="Edit" href={getPageLink("triggerEdit", id)} icon={<EditIcon />}>
+                    <RouterLink
+                        data-tid="Edit"
+                        to={getPageLink("triggerEdit", id)}
+                        icon={<EditIcon />}
+                    >
                         Edit
-                    </Link>
+                    </RouterLink>
                     <span className={cn("control")}>
                         <Tooltip
                             render={() => {
@@ -167,6 +165,7 @@ export default function TriggerInfo({
                             }}
                         >
                             <MaintenanceSelect
+                                icon={<ClockIcon />}
                                 maintenance={maintenance}
                                 caption={maintenanceCaption(delta)}
                                 onSetMaintenance={onSetMaintenance}
@@ -175,8 +174,8 @@ export default function TriggerInfo({
                     </span>
                     <DropdownMenu
                         caption={
-                            <Button use="link">
-                                Other <ArrowTriangleDownIcon color="#6b99d3" />
+                            <Button rightIcon={<ArrowTriangleDownIcon />} use="link">
+                                Other
                             </Button>
                         }
                     >
@@ -269,7 +268,7 @@ export default function TriggerInfo({
                         <dd>
                             <TagGroup
                                 onClick={(tag) => {
-                                    history?.push(
+                                    navigate(
                                         `/?${queryString.stringify(
                                             { tags: [tag] },
                                             {
