@@ -1,9 +1,8 @@
-import React from "react";
-import { useTheme } from "../../../Themes";
+import React, { useMemo } from "react";
 import { createHtmlLegendPlugin } from "../../ContactEventStats/Components/htmlLegendPlugin";
 import { NoisinessDataset } from "../../TriggerNoisiness/TiggerNoisinessChart";
 import { Bar } from "react-chartjs-2";
-import { triggerEventsChartOptions } from "../../../helpers/getChartOptions";
+import { TriggerEventsChartOptions } from "../../../helpers/getChartOptions";
 import { ChartOptions } from "chart.js";
 import EditIcon from "@skbkontur/react-icons/Edit";
 
@@ -11,9 +10,18 @@ export const ContactNoisinessChartView: React.FC<{
     datasets: NoisinessDataset[];
     onEditClick: (label: string) => void;
 }> = ({ datasets, onEditClick }) => {
-    const theme = useTheme();
-
-    const htmlLegendPlugin = createHtmlLegendPlugin(true, null, EditIcon, onEditClick);
+    const htmlLegendPlugin = useMemo(
+        () =>
+            createHtmlLegendPlugin((lable) => (
+                <EditIcon
+                    onClick={(e) => {
+                        e.stopPropagation();
+                        onEditClick(lable);
+                    }}
+                />
+            )),
+        []
+    );
 
     const data = {
         labels: [""],
@@ -27,11 +35,7 @@ export const ContactNoisinessChartView: React.FC<{
                 data={data}
                 plugins={[htmlLegendPlugin]}
                 options={
-                    triggerEventsChartOptions(
-                        "Contacts",
-                        "Number of Events",
-                        theme.chartGridLinesColor
-                    ) as ChartOptions<"bar">
+                    TriggerEventsChartOptions("Contacts", "Number of Events") as ChartOptions<"bar">
                 }
                 style={{ maxHeight: "305px" }}
             />
