@@ -1,7 +1,7 @@
 import React, { useState, useRef } from "react";
 import ArrowTriangleDownIcon from "@skbkontur/react-icons/ArrowTriangleDown";
 import { RenderLayer } from "@skbkontur/react-ui/internal/RenderLayer";
-import { DropdownContainer } from "@skbkontur/react-ui/internal/DropdownContainer";
+import { Popup } from "@skbkontur/react-ui/internal/Popup";
 import { Button } from "@skbkontur/react-ui/components/Button/Button";
 import { MenuItem } from "@skbkontur/react-ui/components/MenuItem";
 import { Menu } from "@skbkontur/react-ui/internal/Menu";
@@ -22,17 +22,24 @@ type MaintenanceSelectProps = {
     caption: React.ReactNode;
     maintenance?: number;
     onSetMaintenance: (maintenance: number) => void;
+    icon?: React.ReactElement;
 };
 
-export default function MaintenanceSelect(props: MaintenanceSelectProps): React.ReactElement {
-    const { caption, maintenance, onSetMaintenance } = props;
+export default function MaintenanceSelect({
+    caption,
+    maintenance,
+    onSetMaintenance,
+    icon,
+}: MaintenanceSelectProps): React.ReactElement {
     const [opened, setOpened] = useState(false);
     const [customMenuShow, setCustomMenuShow] = useState(false);
     const containerEl = useRef(null);
+
     const handleClose = () => {
         setOpened(false);
         setCustomMenuShow(false);
     };
+
     const handleSetMaintenance = (changedMaintenance: number) => {
         handleClose();
         onSetMaintenance(changedMaintenance);
@@ -45,11 +52,13 @@ export default function MaintenanceSelect(props: MaintenanceSelectProps): React.
                     data-tid="TriggerMaintenanceButton"
                     onClick={() => setOpened(true)}
                     use="link"
+                    rightIcon={<ArrowTriangleDownIcon />}
+                    icon={icon}
                 >
-                    {caption} <ArrowTriangleDownIcon color="#6b99d3" />
+                    {caption}
                 </Button>
-                {opened ? (
-                    <DropdownContainer getParent={() => containerEl.current} align="right">
+                {opened && (
+                    <Popup hasShadow anchorElement={containerEl.current} opened={true}>
                         {customMenuShow ? (
                             <CustomMaintenanceMenu
                                 maxDate={lastDayOfMonth(addMonths(new Date(), 1))}
@@ -58,7 +67,7 @@ export default function MaintenanceSelect(props: MaintenanceSelectProps): React.
                                 setMaintenance={handleSetMaintenance}
                             />
                         ) : (
-                            <Menu maxHeight={600}>
+                            <Menu hasMargin={false} maxHeight={600}>
                                 {MaintenanceList.map((maintenance) => (
                                     <MenuItem
                                         key={maintenance}
@@ -74,8 +83,8 @@ export default function MaintenanceSelect(props: MaintenanceSelectProps): React.
                                 <MenuItem onClick={() => setCustomMenuShow(true)}>Custom</MenuItem>
                             </Menu>
                         )}
-                    </DropdownContainer>
-                ) : null}
+                    </Popup>
+                )}
             </span>
         </RenderLayer>
     );
