@@ -1,22 +1,20 @@
 import * as React from "react";
-import { Link, LinkProps } from "react-router-dom";
-import classNames from "classnames/bind";
-
-import styles from "./RouterLink.less";
-
-const cn = classNames.bind(styles);
+import { Link as ReactRouterLink, LinkProps as RouterLinkProps } from "react-router-dom";
+import { Link as KonturLink } from "@skbkontur/react-ui";
+import type { LinkProps as KonturLinkProps } from "@skbkontur/react-ui/components/Link/Link";
 
 type Props = {
-    icon?: React.ReactNode;
-} & LinkProps<typeof Link>;
+    icon?: React.ReactElement | undefined;
+} & Omit<RouterLinkProps, "ref">;
 
-export default function RouterLink(props: Props): React.ReactElement<typeof Link> {
-    const { icon, className, children } = props;
+export default function RouterLink({
+    icon,
+    children,
+    ...rest
+}: Props): React.ReactElement<KonturLinkProps> {
     return (
-        <Link className={cn(className, "link")} {...props}>
-            {icon}
-            {icon ? " " : ""}
-            <span className={cn("content")}>{children}</span>
-        </Link>
+        <KonturLink component={ReactRouterLink} icon={icon} {...rest}>
+            {children}
+        </KonturLink>
     );
 }
