@@ -1,6 +1,5 @@
 import React, { ComponentType } from "react";
-import { hot } from "react-hot-loader/root";
-import { Switch, Route } from "react-router-dom";
+import { useRoutes } from "react-router-dom";
 import MobileErrorContainer from "./Containers/MobileErrorContainer";
 import { getPagePath } from "./Domain/Global";
 import TriggerList, { TriggerListProps } from "./pages/trigger-list/trigger-list";
@@ -13,41 +12,40 @@ import { MobileSettingsPage } from "./Components/Mobile/MobileSettingsPage/Mobil
 import { TeamSettingsPrivateRoute } from "./PrivateRoutes/TeamSettingsPrivateRoute";
 
 type ResponsiveRouteProps = {
-    exact?: boolean;
-    path: string;
     container: ComponentType<TriggerListProps> | ComponentType<TriggerProps>;
     view: ComponentType<TriggerListMobileProps> | ComponentType<TriggerMobileProps>;
 };
 
-function ResponsiveRoute({ container: Container, view: View, ...rest }: ResponsiveRouteProps) {
+function ResponsiveRoute({ container: Container, view: View }: ResponsiveRouteProps) {
     // @ts-ignore problem with typing view
-    return <Route {...rest} render={(props) => <Container {...props} view={View} />} />;
+    return <Container view={View} />;
 }
 
-function Mobile() {
-    return (
-        <Switch>
-            <ResponsiveRoute
-                exact
-                path={getPagePath("index")}
-                container={TriggerList}
-                view={TriggerListMobile}
-            />
-            <ResponsiveRoute
-                exact
-                path={getPagePath("trigger")}
-                container={Trigger}
-                view={TriggerMobile}
-            />
-            <Route exact path={getPagePath("settings")} component={MobileSettingsPage} />
-            <TeamSettingsPrivateRoute
-                exact
-                path={getPagePath("teamSettings")}
-                component={MobileSettingsPage}
-            />
-            <Route component={MobileErrorContainer} />
-        </Switch>
-    );
+function Mobile(): React.ReactElement | null {
+    const routing = useRoutes([
+        {
+            path: getPagePath("index"),
+            element: <ResponsiveRoute container={TriggerList} view={TriggerListMobile} />,
+        },
+        {
+            path: getPagePath("trigger"),
+            element: <ResponsiveRoute container={Trigger} view={TriggerMobile} />,
+        },
+        {
+            path: getPagePath("settings"),
+            element: <MobileSettingsPage />,
+        },
+        {
+            path: getPagePath("teamSettings"),
+            element: <TeamSettingsPrivateRoute Component={MobileSettingsPage} />,
+        },
+        {
+            path: "*",
+            element: <MobileErrorContainer />,
+        },
+    ]);
+
+    return routing;
 }
 
-export default hot(Mobile);
+export default Mobile;
